vfontas: switch vertex to be sorted by y,x
The bitmap is processed top-to-bottom, so new squares only ever
appear at the bottom, which means we can already run the edge removal
at the top, without having to wait for the entire bitmap to be
converted to squares. For that, it is worthwhile to have emap ordered
by rows so that edges above the currently processed row form a
contiguous range in emap.